+++ package/index.html
@@ -5,7 +5,7 @@
         <link
             rel="stylesheet"
             data-name="vs/editor/editor.main"
-            href="../node_modules/monaco-editor-core/dev/vs/editor/editor.main.css"
+            href="./node_modules/monaco-editor-core/dev/vs/editor/editor.main.css"
         />
         <!--Override default .vs .Function style rule, otherwise intellisense is blue.-->
         <style>
@@ -20,7 +20,7 @@
         <script></script>
         <div id="container" style="width: 800px; height: 600px; border: 1px solid grey"></div>
 
-        <script data-name="../node_modules/monaco-editor-core/dev/" src="../node_modules/requirejs/require.js"></script>
+        <script data-name="./node_modules/monaco-editor-core/dev/" src="./node_modules/requirejs/require.js"></script>
         <script src="./test/test.js"></script>
         <div>
             <button onclick="setKuskus()">Kuskus</button>
