+++ package/package.json
@@ -18,7 +18,8 @@
     },
     "scripts": {
         "build": "ts-node ./scripts/build.ts",
-        "watch": "echo dev > test/mode.txt && tsc -p ./tsconfig.watch.json && npm run copy_runtime_deps_to_out && concurrently \"live-server ./ \" \"tsc -w -p ./tsconfig.watch.json \"",
+        "watch": "echo dev > test/mode.txt && concurrently \"live-server ./ \" \"tsc -w -p ./scripts/tsconfig.watch.json \" && \"yarn watch:js\"",
+        "watch:js": "yarn rollup -c ./scripts/rollup.esm.js -w --bundleConfigAsCjs",
         "typecheck": "tsc",
         "test_release": "echo release > test/mode.txt && http-server -c-1 -p 8080 ./ -o index.html",
         "test": "jest",
@@ -39,11 +40,13 @@
         "@rollup/plugin-commonjs": "^24.0.1",
         "@rollup/plugin-node-resolve": "^15.0.1",
         "@rollup/plugin-terser": "^0.4.0",
+        "@rollup/plugin-typescript": "^11.0.0",
         "@tsconfig/node16": "^1.0.3",
         "@types/jest": "^21.1.8",
         "@types/lodash-es": "^4.17.6",
         "@types/node": "16",
         "@types/xregexp": "^4.4.0",
+        "concurrently": "^7.6.0",
         "http-server": "^0.12.3",
         "jest": "^26.1.0",
         "live-server": "^1.2.1",
